+++ src/components/pages/campaigns-create-secure/components/countries-list/types.tsx
@@ -1,0 +1,10 @@
+import { TCountry } from 'types'
+
+export type TOnRemove = (id: string) => void
+
+export type TProps = {
+  data: TCountry[]
+  onRemove: TOnRemove
+  disabled: boolean
+}
+