+++ src/types/campaign-new.tsx
@@ -20,6 +20,7 @@ type TCampaignNew = {
   claim_pattern: TClaimPattern
   sdk: boolean
   proxy_contract_version: string | number
+  available_countries: string[  ]
   available_wallets: string[] 
 }