+++ src/types/index.js
@@ -57,11 +57,23 @@ import {
   TCollection,
   TCollectionToken
 } from './collection'
+import  {
+  TZerionERC20Item,
+  TZerionERC20ItemRelationships,
+  TZerionERC20ItemAttributes
+} from './zerion-erc20-item'
+import { TZerionNetworkItem } from './zerion-network-item'
+import TCountry from './country'
 
 export {
   TCampaign,
+  TZerionNetworkItem,
   TDispenserStats,
+  TCountry,
   TMnemonicContract,
+  TZerionERC20Item,
+  TZerionERC20ItemRelationships,
+  TZerionERC20ItemAttributes,
   TCampaignCreateStep,
   TFeatureName,
   TDispenserWhitelistType,