+++ src/types/zerion-network-item.tsx
@@ -1,0 +1,5 @@
+export type TZerionNetworkItemAttributes = {
+  id: string
+}
+
+export type TZerionNetworkItem = TZerionNetworkItemAttributes